+++ backend/functions/generateHexColor.js
