+++ backend/functions/serverStatuses.js
@@ -1,0 +1,8 @@
+let serverStatuses = {
+    loopQueue: true,
+    randomizeQueue: false,
+    playState: false,
+    filters: [],
+  };
+
+  module.exports = serverStatuses;
