+++ backend/index.js
@@ -1,78 +1,50 @@
-const express      = require('express')
-const http         = require("http")
-const ytdlp        = require('yt-dlp-exec');
-const cors         = require('cors');
-const ffmpeg       = require('fluent-ffmpeg');
-const ffmpegPath   = require('@ffmpeg-installer/ffmpeg').path;
-const WebSocket    = require('ws');
-const bcrypt       = require('bcrypt');
-const bodyParser   = require('body-parser');
-const jwt          = require('jsonwebtoken'); // Optional
-const authenticateToken = require('./middleware/authenticateToken'); // Middleware
-const SECRET_KEY  = require('./middleware/SECRET_KEY'); // Secret key for JWT
-const generateHexColor = require('./additional functions/generateHexColor');
+const cors          = require('cors');
+const express       = require('express');
+const http          = require('http');
+const WebSocket     = require('ws');
+const router        = express.Router();
 
-let serverStatuses = {
-  loopQueue: true,
-  randomizeQueue: false,
-  playState: false,
-  filters: [],
-};
+const app           = express();
+const server        = http.createServer(app);
+const wss           = new WebSocket.Server({ noServer: true });
+const bodyParser    = require('body-parser');
+
+app.use(cors());
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+app.use(bodyParser.json());
+
+// Load route files
+app.use(require('./routes/userMgmt.js'));
+
+// External functions
+
+const { authenticateToken } = require('./middleware/authenticateToken.js'); // Middleware
+const { broadcastMessage, sendFetchNotification } = require('./functions/websocket.js');
+const { initializeDatabase, changeStatuses, dbQueue, dbTagColors } = require('./functions/database.js');
+const { processTags } = require('./functions/processTags.js');
+const serverStatuses = require('./functions/serverStatuses.js');
 
 let shuffledQueue = [];
 
-const sqlite3      = require('sqlite3').verbose();
-const app          = express()
-const server       = http.createServer(app);
-const wss          = new WebSocket.Server({ server });
 
-const db           = new sqlite3.Database('./queue.db');
-const dbUsers      = new sqlite3.Database('./users.db');
-const dbTagColors  = new sqlite3.Database('./tagColors.db');
+
 const PORT         = process.env.PORT || 5000
 
-
-ffmpeg.setFfmpegPath(ffmpegPath);
-app.use(cors());
-app.use(express.json());
-app.use(bodyParser.json());
-
-// Create queue table if it doesn't exist
-db.serialize(() => {
-  db.run(`CREATE TABLE IF NOT EXISTS queue (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    url TEXT NOT NULL,
-    title TEXT,
-    status TEXT DEFAULT 'pending',
-    audioUrl TEXT,
-    duration INTEGER,
-    tags TEXT
-  )`);
-});
-
-dbUsers.serialize(() => {
-  dbUsers.run(`CREATE TABLE IF NOT EXISTS users (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    username TEXT NOT NULL,
-    pwdHash TEXT NOT NULL
-  )`, (err) => {
-    if (err) {
-        console.error('Error creating table:', err.message);
-    };
+// Handle WebSocket upgrade
+server.on('upgrade', (request, socket, head) => {
+  wss.handleUpgrade(request, socket, head, (ws) => {
+      wss.emit('connection', ws, request);
   });
 });
+require('./functions/websocket.js')(wss); // Import WebSocket logic and pass `wss`
 
-dbTagColors.serialize(() => {
-  dbTagColors.run(`CREATE TABLE IF NOT EXISTS tagColors (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    tag TEXT NOT NULL,
-    color TEXT NOT NULL
-  )`, (err) => {
-    if (err) {
-        console.error('Error creating table:', err.message);
-    };
-  });
-});
+
+initializeDatabase();
+
+// Change all statuses to 'pending' on server start
+changeStatuses('playing', 'pending');
+changeStatuses('paused', 'pending');
 
 class currentAudio {
   constructor({url, index = 0, duration = null}){
@@ -116,173 +88,14 @@ class currentAudio {
 }
 var _currentAudio = new currentAudio({url: null})
 
-// ------------------------------------------------------WEBSOCKET---------------------------------------------------------------
-const activeConnections = new Set();
-
-wss.on('connection', (ws) => {
-  console.log('\x1b[32m%s\x1b[0m', 'Client connected');
-  activeConnections.add(ws);
-
-  wss.emit('statusUpdate', serverStatuses); // Send all statuses to the client
-
-  wss.on('statusUpdate', (updates) => {
-    Object.assign(serverStatuses, updates);  // Merge updates into server state
-    broadcastMessage({ event: 'statusUpdate', message: serverStatuses });
-  });
-
-  if (activeConnections.size > 0){
-    play();
-  }
-
-  ws.on('message', (message) => {
-    // ws.send({message});
-    const messageString = Buffer.from(message).toString();  // Decode buffer to string
-    console.log('Received:', {message, messageString});    
-    const data = JSON.parse(messageString);  // Parse the JSON message
-    switch(data.type){
-
-      case 'params':
-        loopQueue = data.message.loopQueue;
-        randomizeQueue = data.message.randomizeQueue;
-        console.log(data.message.status);
-        console.log(data);
-        break;
-
-      case 'statusUpdate':
-        Object.assign(serverStatuses, data.status); // Update statuses
-        broadcastMessage({ event: 'statusUpdate', message: serverStatuses });
-        break;
-
-      default:
-        console.log('Unhandled data type:', data.type);
-        break;
-    }
-  });
-
-  ws.on('close', () => {
-    console.log('\x1b[35m%s\x1b[0m', 'Client disconnected');
-    activeConnections.delete(ws);
-  });
-});
-
-// Function to send a message to all WebSocket clients
-const broadcastMessage = (message) => {
-  activeConnections.forEach(client => {
-      if (client.readyState === WebSocket.OPEN) {
-          client.send(JSON.stringify(message));
-          console.log('\x1b[34m%s\x1b[0m', `Sent msg: ${JSON.stringify(message).slice(0,1500)}`);
-      }
-  });
-};
-
-// Notify all active WebSocket clients to fetch queue
-const sendFetchNotification = () =>{
-  broadcastMessage({ event: 'refresh', message: JSON.stringify('Refreshing...') });
-}
-// ------------------------------------------------------------------------------------------------------------
-
-
-
 
 // ---------------------------ENDPOINTS----------------------------------
-// ------USERS-------
-
-// Login endpoint
-app.post('/api/login', (req, res) => {
-  const { username, password } = req.body;
-
-  if (!username || !password) {
-      return res.status(400).json({ error: 'Username and password are required.' });
-  }
-
-  // Query the database for the user
-  const sql = `SELECT * FROM users WHERE username = ?`;
-  dbUsers.get(sql, [username], async (err, row) => {
-      if (err) {
-          console.error('Database error:', err.message);
-          return res.status(500).json({ error: 'Internal server error' });
-      }
-
-      if (!row) {
-          return res.status(404).json({ error: 'User not found' });
-      }
-
-      // Compare the provided password with the stored hash
-      const passwordMatch = await bcrypt.compare(password, row.pwdHash);
-      if (!passwordMatch) {
-          return res.status(401).json({ error: 'Invalid credentials' });
-      }
-
-      // Generate a JWT token (optional)
-      const token = jwt.sign({ id: row.id, username: row.username }, SECRET_KEY, { expiresIn: '1h' });
-
-      return res.json({ message: 'Login successful', token });
-  });
-});
-
-// Register endpoint
-app.post('/api/register', async (req, res) => {
-  const { username, password, secret } = req.body;
-  console.log(req.body);
-
-  if (!username || !password) {
-      return res.status(400).json({ error: 'Username and password are required.' });
-  }
-  if (!secret || secret !== '1111') { //completly secure way of authorizing administrator
-    return res.status(400).json({ error: 'You are not authorized to add users.' });
-  }
-
-  try {
-      // Hash the password
-      const saltRounds = 10;
-      const pwdHash = await bcrypt.hash(password, saltRounds);
-
-      // Insert the user into the database
-      const sql = `INSERT INTO users (username, pwdHash) VALUES (?, ?)`;
-      dbUsers.run(sql, [username, pwdHash], function (err) {
-          if (err) {
-              console.error('Error inserting user:', err.message);
-              return res.status(500).json({ error: 'Internal server error' });
-          }
-
-          res.json({ message: 'User registered successfully', userId: this.lastID });
-      });
-  } catch (error) {
-      console.error('Error hashing password:', error.message);
-      res.status(500).json({ error: 'Internal server error' });
-  }
-});
-
-app.get('/api/userinfo', (req, res) => {
-  const token = req.headers.authorization?.split(' ')[1];
-  if (!token) {
-      return res.status(401).json({ error: 'Unauthorized' });
-  }
-
-  try {
-      const decoded = jwt.verify(token, SECRET_KEY);
-      res.json({ id: decoded.id, username: decoded.username });
-  } catch (err) {
-      res.status(403).json({ error: 'Invalid token' });
-  }
-});
-
-
 
 // ------QUEUE-------
 
-// Get audio stream URL
-app.post('/process', async (req, res) => {
-  console.log('body:', req.body)
-  const { url } = req.body;
-  getAudioUrlAndTitle(url).then(({ title, audioUrl }) =>{
-    res.json({ title: title, url: audioUrl })
-  });
-});
-
 // Get the Current Queue
 app.get('/queue', (req, res) => {
-  db.all(`SELECT * FROM queue ORDER BY id`, (err, rows) => {
+  dbQueue.all(`SELECT * FROM queue ORDER BY id`, (err, rows) => {
       if (err) {
       return res.status(500).json({ error: 'Failed to retrieve queue' });
       }
@@ -303,50 +116,31 @@ app.post('/queue/add', authenticateToken, (req, res) => {
     return res.status(400).json({ error: 'YouTube URL is required' });
   }
 
-  const tagsExtended = tags ? [req.user.username, ...tags] : [req.user.username]; // Add the user's username as a tag (at least username)
-  const tagsJson = JSON.stringify(tagsExtended); // Convert tags array to JSON string
+  let tagsJson = processTags(req, tags); //Add tags to db and assign random colors to them
 
-  db.run(`INSERT INTO queue (url, status, tags) VALUES (?, 'pending', ?)`, [url, tagsJson], function (err) {
+  dbQueue.run(`INSERT INTO queue (url, status, tags) VALUES (?, 'pending', ?)`, [url, tagsJson], function (err) {
     if (err) {
       return res.status(500).json({ error: 'Failed to add to queue' });
     }
     res.json({ message: 'URL added to queue', id: this.lastID });
 
-    // Process each tag in tagsExtended
-    tagsExtended.forEach(tag => {
-      dbTagColors.get(`SELECT * FROM tagColors WHERE tag = ?`, [tag], (err, row) => {
-          if (err) {
-              console.error(`Error checking tag: ${tag}`, err.message);
-              return;
-          }
-
-          // If the tag does not exist, insert it with a new color
-          if (!row) {
-            dbTagColors.run(`INSERT INTO tagColors (tag, color) VALUES (?, ?)`, [tag, generateHexColor()], (err) => {
-                  if (err) {
-                      console.error(`Error inserting tag: ${tag}`, err.message);
-                  } else {
-                      // console.log(`Tag '${tag}' added with color assigned.`);
-                  }
-              });
-          }
-      });
-    });
-
-    processNextFromQueue(); // trigger queue processing
+    // processNextFromQueue(serverStatuses.filters); // trigger queue processing
   });
 });
 
+
+
 // Skip endpoint
 app.post('/queue/skip', async (req, res) => {
-  skipTrack(res); // Skip the current track
+  // skipTrack(res); // Skip the current track
+  //TODO
   res.status(200).send('Track skipped');
 });
 
 
 // Clear Queue
 app.post('/queue/clear', authenticateToken, (req, res) => {
-  db.run(`DELETE FROM queue`, (err) => {
+  dbQueue.run(`DELETE FROM queue`, (err) => {
     if (err) return res.status(500).json({ error: 'Failed to clear the queue' });
     serverStatuses.playState = "pause";
     stopPlaying();
@@ -358,7 +152,7 @@ app.post('/queue/clear', authenticateToken, (req, res) => {
 // Remove track from the queue
 app.post('/queue/remove', authenticateToken, (req, res) => {
   const IdToRemove = req.body.id;
-  db.run(`DELETE FROM queue WHERE id=?`, [IdToRemove], (err) => {
+  dbQueue.run(`DELETE FROM queue WHERE id=?`, [IdToRemove], (err) => {
     if (err) return res.status(500).json({ error: 'Failed to remove from the queue' });
     sendFetchNotification();
     res.json({ message: 'Track removed from the queue' });
@@ -367,7 +161,7 @@ app.post('/queue/remove', authenticateToken, (req, res) => {
 
 // Pause track
 app.post('/queue/pause', (req, res) => {
-  db.run(`UPDATE queue SET status = 'paused' WHERE status = 'playing'`, (err) => {
+  dbQueue.run(`UPDATE queue SET status = 'paused' WHERE status = 'playing'`, (err) => {
     if (err) return res.status(500).json({ error: 'Failed to pause the track' });
     _currentAudio.stopIndexIncrement();
     stopPlaying();
@@ -378,7 +172,7 @@ app.post('/queue/pause', (req, res) => {
 
 // Resume track
 app.post('/queue/resume', (req, res) => {
-  db.run(`UPDATE queue SET status = 'playing' WHERE status = 'paused'`, (err) => {
+  dbQueue.run(`UPDATE queue SET status = 'playing' WHERE status = 'paused'`, (err) => {
     if (err) return res.status(500).json({ error: 'Failed to resume the track' });
     play();
     sendFetchNotification();
@@ -407,280 +201,11 @@ app.get('/queue/tags', (req, res) => {
 
 // ------------------------------------------------------FUNCTIONS---------------------------------------------------------------
 
-const getDuration = async (url) => {
-  try {
-    // Use yt-dlp to get the duration
-    const result = await ytdlp(url, {
-        dumpSingleJson: true, // Outputs metadata in JSON format
-    });
-    // Check if the URL has a "dur" parameter
-    const urlObj = new URL(result.url || ''); // Ensure result.url is valid
-    const durationParam = urlObj.searchParams.get('dur'); // Get "dur" from the URL
-
-    if (!durationParam) {
-      throw new Error('Duration parameter not found in the URL');
-    }
-
-    const durationInSeconds = parseInt(durationParam); // Convert to seconds
-
-    // Respond with the duration
-    //(new Date(durationInSeconds * 1000).toISOString().substring(11, 8)) // Format to hh:mm:ss
-    return(durationInSeconds);
-  } catch (error) {
-      console.error('Error fetching audio length:', error);
-      return(null);
-  }
-}
-
-// Function to get the best audio stream URL using yt-dlp
-const getAudioUrlAndTitle = async (url) => {
-  try {
-    // Execute yt-dlp and fetch the required outputs
-    const output = await ytdlp(url, {
-      f: 'bestaudio', // Fetch the best audio format
-      g: true,        // Get the direct audio URL
-      print: 'title'  // Print the video title
-    });
-
-    // Parse the output
-    const lines = output.split('\n').filter(line => line.trim() !== '');
-    const videoTitle = lines[0]; // First line is the title
-    const audioUrl = lines[1];   // Second line is the audio URL
-
-    return { title: videoTitle, audioUrl: audioUrl };
-  } catch (error) {
-    console.error(`yt-dlp error: ${error}`);
-    throw new Error('Failed to fetch audio stream URL');
-  }
-};
-
-
-
-// Helper function to fetch all pending items based on tags
-const getAllPendingItems = async (tags = []) => {
-  let query = `SELECT * FROM queue WHERE status = 'pending'`;
-  const params = [];
-
-  // Add tag filtering if tags are provided
-  if (tags.length > 0) {
-    const tagConditions = tags.map(() => `tags LIKE ?`).join(' AND ');
-    query += ` AND (${tagConditions})`;
-    params.push(...tags.map(tag => `%${tag}%`)); // Use LIKE for partial match
-  }
-
-  return new Promise((resolve, reject) => {
-    db.all(query, params, (err, items) => {
-      if (err) {
-        console.error('[getAllPendingItems] Error fetching items:', err);
-        reject(err);
-      } else {
-        resolve(items || []);
-      }
-    });
-  });
-};
-
-// Helper function to shuffle an array
-const shuffleArray = (array) => {
-  for (let i = array.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-  return array;
-};
-
-// Helper function to fetch the next item (non-shuffle mode)
-const getNextItem = async (tags = []) => {
-  let query = `SELECT * FROM queue WHERE status = 'pending'`;
-  const params = [];
-
-  // Add tag filtering if tags are provided
-  if (tags.length > 0) {
-    const tagConditions = tags.map(() => `tags LIKE ?`).join(' AND ');
-    query += ` AND (${tagConditions})`;
-    params.push(...tags.map(tag => `%${tag}%`)); // Use LIKE for partial match
-  }
-
-  // Default ordering
-  query += ` ORDER BY id LIMIT 1`;
-
-  return new Promise((resolve, reject) => {
-    db.get(query, params, (err, item) => {
-      if (err) {
-        console.error('[getNextItem] Error fetching item:', err);
-        reject(err);
-      } else {
-        resolve(item || null);
-      }
-    });
-  });
-};
-
-// Helper function to process and play a song
-const playSong = async (item) => {
-  try {
-    // Mark the item as processing
-    await updateItemStatus(item.id, 'processing');
-    sendFetchNotification();
-
-    // Fetch and process the audio URL and title
-    const { title, audioUrl } = await getAudioUrlAndTitle(item.url);
-    const duration = await getDuration(audioUrl);
-
-    // Update the database with the processed item
-    db.run(
-      `UPDATE queue SET status = 'processed', audioUrl = ?, title = ?, duration = ? WHERE id = ?`,
-      [audioUrl, title, duration, item.id],
-      (err) => {
-        if (err) {
-          console.error('[playSong] Error updating processed item:', err);
-        } else {
-          console.log(`[playSong] Successfully processed item: ${title}`);
-          sendFetchNotification();
-          play();
-        }
-      }
-    );
-  } catch (error) {
-    console.error('[playSong] Error:', error.message);
-  }
-};
-
-// Helper function to update item status
-const updateItemStatus = async (id, status) => {
-  return new Promise((resolve, reject) => {
-    db.run(`UPDATE queue SET status = ? WHERE id = ?`, [status, id], (err) => {
-      if (err) {
-        console.error(`[updateItemStatus] Error updating status to ${status}:`, err);
-        reject(err);
-      } else {
-        resolve();
-      }
-    });
-  });
-};
-
-// Main processing function
-const processNextFromQueue = async (tags = []) => {
-  try {
-    const randomizeQueue = serverStatuses.randomizeQueue;
-
-    // Shuffle mode logic
-    if (randomizeQueue) {
-      // Initialize or reset the shuffled queue if it's empty
-      if (!shuffledQueue || shuffledQueue.length === 0) {
-        console.log('[processNextFromQueue] Shuffling the queue...');
-        await resetFinishedToPending(); // Reset all finished items to pending
-        const allPendingItems = await getAllPendingItems(tags);
-
-        if (allPendingItems.length === 0) {
-          console.log('[processNextFromQueue] No pending items found for shuffle.');
-          return null;
-        }
-
-        shuffledQueue = shuffleArray(allPendingItems); // Shuffle and store
-      }
-
-      // Get the next song from the shuffled queue
-      const item = shuffledQueue.shift(); // Remove and get the first song
-      await playSong(item); // Process the song
-      return true;
-    }
-
-    // Normal queue logic (no randomization)
-    let item = await getNextItem(tags);
-
-    if (!item) {
-      console.log('[processNextFromQueue] No pending items found.');
-
-      // Check if the queue has finished items and loop if necessary
-      const lastQueueStatus = await getLastQueueStatus();
-      if (lastQueueStatus && lastQueueStatus.status === 'finished' && serverStatuses.loopQueue) {
-        console.log('[processNextFromQueue] Looping the queue.');
-        await resetFinishedToPending();
-        item = await getNextItem(tags); // Try fetching again after reset
-        if (!item) {
-          console.log('[processNextFromQueue] No items to process after loop reset.');
-          return null;
-        }
-      } else {
-        console.log('[processNextFromQueue] Queue is not finished or looping is disabled.');
-        return null;
-      }
-    }
-
-    // Process the item normally
-    await playSong(item);
-    return true;
-  } catch (error) {
-    console.error('[processNextFromQueue] Error:', error.message);
-    return null;
-  }
-};
-
-// Helper function to fetch the last queue status
-const getLastQueueStatus = async () => {
-  return new Promise((resolve, reject) => {
-    db.get(`SELECT status FROM queue ORDER BY id DESC LIMIT 1`, [], (err, row) => {
-      if (err) {
-        console.error('[getLastQueueStatus] Error accessing queue:', err);
-        reject(err);
-      } else {
-        resolve(row || null);
-      }
-    });
-  });
-};
-
-// Helper function to reset finished items to pending
-const resetFinishedToPending = async () => {
-  return new Promise((resolve, reject) => {
-    db.run(`UPDATE queue SET status = 'pending' WHERE status = 'finished'`, (err) => {
-      if (err) {
-        console.error('[resetFinishedToPending] Error updating status:', err);
-        reject(err);
-      } else {
-        console.log('[resetFinishedToPending] Status updated to pending for all finished songs.');
-        resolve();
-      }
-    });
-  });
-};
-
-
-
-
-// Get new audioUrl for row with certain id
-const reprocessRowById = async (id) => {
-  try {
-    // Mark it as processing
-    db.run(`UPDATE queue SET status = 'processing' WHERE id = ?`, [id]);
-    sendFetchNotification();
-
-    // Get the URL to reprocess
-    const itemToReprocess = await getById(id);
-    const urlToReprocess = itemToReprocess.url;
-    console.log(`[reprocessRowById] Reprocessing url: ${urlToReprocess}`);
-
-    // Run yt-dlp to get the audio URL
-    const { title, audioUrl } = await getAudioUrlAndTitle(urlToReprocess);
-
-    // Update the database with the processed status and audio URL
-    db.run(`UPDATE queue SET status = 'processed', audioUrl = ?, title = ?  WHERE id = ?`, [audioUrl, title, id]);
-    sendFetchNotification();
-
-    return true; // Resolves with true
-  } catch (error) {
-    console.error('Error processing row:', error);
-    throw error; // Propagates the error
-  }
-};
-
 
 // returns object item where status = 'playing'
 const currentlyPlaying = () => {
   return new Promise((resolve, reject) => {
-    db.get(`SELECT * FROM queue WHERE status = 'playing' ORDER BY id LIMIT 1`, (err, item) => {
+    dbQueue.get(`SELECT * FROM queue WHERE status = 'playing' ORDER BY id LIMIT 1`, (err, item) => {
       if (err) {
         console.error('[currentlyPlaying] Error accessing queue:', err);
         reject(err); // Reject the promise with the error
@@ -688,25 +213,8 @@ const currentlyPlaying = () => {
       }
       // NOTHING PLAYING = FETCHING NEXT PROCESSED
       if (!item) {
-        console.log('[currentlyPlaying] Fetching next processed track.');
-        // Fetching next processed entry and marking it as playing
-        db.get(`SELECT * FROM queue WHERE status = 'processed' ORDER BY id LIMIT 1`, (err, item) => {
-          if (err) {
-            console.error('[currentlyPlaying] Error accessing queue:', err);
-            reject(err); // Reject the promise with the error
-            return;
-          }
-          // NOTHING PROCESSED = RETURN NULL
-          if (!item) {
-            console.log('[currentlyPlaying] No processed songs.');
-            resolve(null);
-            return;
-          }
-          db.run(`UPDATE queue SET status = 'playing' WHERE id = ?`, [item.id]);
-          play();
-          sendFetchNotification();
-          resolve(JSON.stringify(item));
-        });
+        console.log('[currentlyPlaying] Nothing with status "playing".');
+        
         resolve(null); // Resolve with null if no playing or processed songs
         return;
       }
@@ -722,7 +230,7 @@ const play = async () => {
     let currPlaying = await currentlyPlaying();
     if (currPlaying == null){
       console.log(`[play] currently playing (null): ${currPlaying}`);
-      await processNextFromQueue();
+      await processNextFromQueue(serverStatuses.filters);
     }else{
       serverStatuses.playState = "play";
       _currentAudio.duration = JSON.parse(currPlaying).duration;
@@ -740,7 +248,7 @@ const next = async () => {
 
 const stopPlaying = async () => {
   return new Promise((resolve, reject) => {
-    db.all(`SELECT * FROM queue WHERE status = 'playing' ORDER BY id`, (err, items) => {
+    dbQueue.all(`SELECT * FROM queue WHERE status = 'playing' ORDER BY id`, (err, items) => {
       if (err) {
         console.error('[stopPlaying] Error accessing queue:', err);
         console.log('[stopPlaying] Retrying...');
@@ -754,7 +262,7 @@ const stopPlaying = async () => {
         return;
       }
       items.forEach((item) => {
-        db.run(`UPDATE queue SET status = 'finished' WHERE id = ?`, [item.id]);
+        dbQueue.run(`UPDATE queue SET status = 'finished' WHERE id = ?`, [item.id]);
         console.log(`[stopPlaying] Finished item with id=${item.id}`);
         // broadcastMessage({ event: 'track-ended', message: 0 });
         serverStatuses.playState = "pause";
@@ -766,44 +274,8 @@ const stopPlaying = async () => {
   });
 }
 
-// Find url with given id of the row
-const getById = (id) => {
-  return new Promise((resolve, reject) => {
-    db.get(`SELECT * FROM queue WHERE id = ? ORDER BY id LIMIT 1`, [id], (err, item) => {
-      if (err) {
-        console.error('[getUrlById] Error accessing db: ', err);
-        reject(err); // Reject the promise with the error
-        return;
-      }
-      if (!item) {
-        console.error(`[getUrlById] No data with id =${id}: `, err);
-        reject(null); // Resolve with null
-        return;
-      }
-      console.log(`[getUrlById] Found row with id=${id}`)
-      resolve(item); // Resolve with URL
-    });
-  });
-};
 
-const isQueueEmpty = () => {
-  return new Promise((resolve, reject) => {
-    db.get(`SELECT * FROM queue WHERE NOT status = 'finished' ORDER BY id LIMIT 1`, (err, item) => {
-      if (err) {
-        console.error('[isQueueEmpty] Error accessing queue:', err);
-        resolve(null);
-        return;
-      }
-      if (!item) {
-        console.log('[isQueueEmpty] No processed songs.');
-        resolve(true);
-        return;
-      }
-      resolve(false);
-      return;
-    });
-  });
-}
+
 
 // shorten links from error messages
 function formatError(err) {
@@ -822,3 +294,4 @@ function formatError(err) {
 
 
 server.listen(PORT, () => console.log('\x1b[2m%s\x1b[0m', `Running server on port ${PORT}`))
+module.exports = app;
