+++ backend/middleware/authenticateToken.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken');
-const SECRET_KEY = require('./SECRET_KEY');
+const SECRET_KEY = require('./SECRET_KEY.js');
 
 const authenticateToken = (req, res, next) => {
     const authHeader = req.headers.authorization;
